Correct data type for voltage resolution

+++ src/MQUnifiedsensor.h
@@ -276,7 +276,7 @@ class MQUnifiedsensor
     int _pin, _type, _lecturePosInArray; 
     double  _R0;
     String  _nameLectureRequeired;   
-    int _VOLT_RESOLUTION  = 5.0; // if 3.3v use 3.3
+    float _VOLT_RESOLUTION  = 5.0; // if 3.3v use 3.3
     int _RLValue = 10; //Value in KiloOhms
     float _PPM, _RS_Calc;
     float _ratioInCleanAir, _sensor_volt, RS_air, _a, _b, _ratio;
